+++ assets/javascripts/kitten/karl/buttons/payment-button.js
@@ -4,6 +4,7 @@ import { TagButton } from 'kitten/components/buttons/tag-button'
 import { VisaIcon } from 'kitten/components/icons/visa-icon'
 import { MasterCardIcon } from 'kitten/components/icons/mastercard-icon'
 import { CbIcon } from 'kitten/components/icons/cb-icon'
+import { Line } from 'kitten/components/layout/line'
 
 const paymentButtonWrapper = (WrappedComponent, props) => {
   return class PaymentButtonWrapper extends React.Component {
@@ -20,20 +21,37 @@ const paymentButtonWrapper = (WrappedComponent, props) => {
   }
 }
 
-export const KarlPaymentButtonVisa = paymentButtonWrapper(VisaIcon, {
+export const KarlPaymentButtons = () => (
+  <Line>
+    <Line.Item>
+      <KarlPaymentButtonVisa />
+    </Line.Item>
+    <Line.Item>
+      <KarlPaymentButtonMasterCard />
+    </Line.Item>
+    <Line.Item>
+      <KarlPaymentButtonCb />
+    </Line.Item>
+    <Line.Item>
+      <KarlPaymentButtonBankTransfer />
+    </Line.Item>
+  </Line>
+)
+
+const KarlPaymentButtonVisa = paymentButtonWrapper(VisaIcon, {
   name: 'payment-button-1',
 })
 
-export const KarlPaymentButtonMasterCard = paymentButtonWrapper(MasterCardIcon, {
+const KarlPaymentButtonMasterCard = paymentButtonWrapper(MasterCardIcon, {
   name: 'payment-button-1',
   defaultChecked: true,
 })
 
-export const KarlPaymentButtonCb = paymentButtonWrapper(CbIcon, {
+const KarlPaymentButtonCb = paymentButtonWrapper(CbIcon, {
   name: 'payment-button-1',
 })
 
-export const KarlPaymentButtonBankTransfer = props => {
+const KarlPaymentButtonBankTransfer = props => {
   const text = <TagButton big tag="span">
                  Virement
                </TagButton>