+++ assets/javascripts/kitten/components/cards/project-card.js
@@ -41,7 +41,11 @@ class ProjectCardComponent extends React.Component {
         </Marger>
 
         <Marger top="1" bottom="1.5">
-          <Title margin={ false } modifier="quaternary">{ title }</Title>
+          <Title margin={ false }
+                 modifier="quaternary"
+                 tag="p">
+            { title }
+          </Title>
         </Marger>
       </div>
     )
@@ -140,6 +144,7 @@ class ProjectCardComponent extends React.Component {
       statusTertiaryBackground,
       statusGreyBackground,
       statusErrorBackground,
+      statusErrorReverseBackground,
       statusWithoutTopBorder,
     } = this.props
 
@@ -154,6 +159,8 @@ class ProjectCardComponent extends React.Component {
         'k-ProjectCard__status--tertiaryBackground': statusTertiaryBackground,
         'k-ProjectCard__status--greyBackground': statusGreyBackground,
         'k-ProjectCard__status--errorBackground': statusErrorBackground,
+        'k-ProjectCard__status--errorReverseBackground':
+          statusErrorReverseBackground,
       },
     )
 
@@ -216,6 +223,7 @@ ProjectCardComponent.defaultProps = {
   statusTertiaryBackground: false,
   statusGreyBackground: false,
   statusErrorBackground: false,
+  statusErrorReverseBackground: false,
   statusWithoutTopBorder: false,
 }