+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Add new `MasterCardIcon` and `CbIcon` components.
+- Feature: Add `.k-TagButton__icon` class to adds icons to tag buttons.
+
 ## [11.4.0] - 2017-05-04
 
 Feature: