+++ CHANGELOG.md
@@ -4,6 +4,16 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+## [11.4.0] - 2017-05-04
+
+Feature:
+- Add status in `ProjectCard`.
+
+Fixes:
+- Fix `LegoGrid` grid.
+- Fix tag `Title` in `ProjectCardComponent`.
+- Fix test launch.
+
 ## [11.3.0] - 2017-05-02
 
 Features: